+++ examples/xbr/test_typed.py
@@ -8,50 +8,50 @@ from autobahn import xbr
 
 import eth_keys
 from eth_account import Account
-
+from ethereum import utils
 from crossbarfx.cfxdb import pack_uint256, unpack_uint256, pack_uint128, unpack_uint128
 
-
-data1 = {
-    'types': {
-        'EIP712Domain': [
-            {'name': 'name', 'type': 'string'},
-            {'name': 'version', 'type': 'string'},
-            {'name': 'chainId', 'type': 'uint256'},
-            {'name': 'verifyingContract', 'type': 'address'},
+# https://lib.rs/crates/eip-712
+# hash: be609aee343fb3c4b28e1df9e632fca64fcfaede20f02e86244efddf30957bd2
+data2 = {
+    "primaryType": "Mail",
+    "domain": {
+        "name": "Ether Mail",
+        "version": "1",
+        "chainId": 1,
+        "verifyingContract": "0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC"
+    },
+    "message": {
+        "from": {
+            "name": "Cow",
+            "wallet": "0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826"
+        },
+        "to": {
+            "name": "Bob",
+            "wallet": "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB"
+        },
+        "contents": "Hello, Bob!"
+    },
+    "types": {
+        "EIP712Domain": [
+            {"name": "name", "type": "string"},
+            {"name": "version", "type": "string"},
+            {"name": "chainId", "type": "uint256"},
+            {"name": "verifyingContract", "type": "address"}
         ],
-        'Person': [
-            {'name': 'name', 'type': 'string'},
-            {'name': 'wallet', 'type': 'address'}
+        "Person": [
+            {"name": "name", "type": "string"},
+            {"name": "wallet", "type": "address"}
         ],
-        'Mail': [
-            {'name': 'from', 'type': 'Person'},
-            {'name': 'to', 'type': 'Person'},
-            {'name': 'contents', 'type': 'string'}
+        "Mail": [
+            {"name": "from", "type": "Person"},
+            {"name": "to", "type": "Person"},
+            {"name": "contents", "type": "string"}
         ]
-    },
-    'primaryType': 'Mail',
-    'domain': {
-        'name': 'Ether Mail',
-        'version': '1',
-        'chainId': 1,
-        'verifyingContract': '0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC',
-    },
-    'message': {
-        'from': {
-            'name': 'Cow',
-            'wallet': '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
-        },
-        'to': {
-            'name': 'Bob',
-            'wallet': '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
-            'foobar': 23,
-        },
-        'contents': 'Hello, Bob!',
-    },
+    }
 }
 
-data2 = {
+data = {
     'types': {
         'EIP712Domain': [
             {'name': 'name', 'type': 'string'},
@@ -59,53 +59,39 @@ data2 = {
             {'name': 'chainId', 'type': 'uint256'},
             {'name': 'verifyingContract', 'type': 'address'},
         ],
-        'Transaction': [
-            # The buyer delegate Ethereum address. The technical buyer is usually the
-            # XBR delegate of the XBR consumer/buyer of the data being bought.
-            {'name': 'buyer_adr', 'type': 'address'},
-
-            # The buyer delegate Ed25519 public key.
-            {'name': 'buyer_pubkey', 'type': 'uint256'},
-
-            # The UUID of the data encryption key to buy.
-            {'name': 'key_id', 'type': 'uint128'},
-
-            # Amount signed off to pay. The actual amount paid is always less than or
-            # equal to this, but the amount must be greater than or equal to the price in the
-            # offer for selling the data encryption key being bought.
-            {'name': 'amount', 'type': 'uint256'},
-
-            # Amount remaining in the payment channel after the transaction.
+        'ChannelClose': [
+            {'name': 'channel_adr', 'type': 'address'},
+            {'name': 'channel_seq', 'type': 'uint32'},
             {'name': 'balance', 'type': 'uint256'},
         ],
     },
-    'primaryType': 'Transaction',
+    'primaryType': 'ChannelClose',
     'domain': {
         'name': 'XBR',
         'version': '1',
-
-        # test chain/network ID
-        'chainId': 5777,
-
-        # XBRNetwork contract address
-        'verifyingContract': '0x254dffcd3277c0b1660f6d42efbb754edababc2b',
+        'chainId': 1,
+        'verifyingContract': '0x254dffcd3277C0b1660F6d42EFbB754edaBAbC2B',
     },
-    'message': {
-        'buyer_adr': '0x78Abb38526c7F70d10EBcDf77941B61f193856f5',
-        'buyer_pubkey': unpack_uint256(a2b_hex('ebdfef6d225155873355bd4afeb2ed3100b0e0b5fddad12bd3cd498c1e0c1fbd')),
-        'key_id': unpack_uint128(a2b_hex('c37ba03c32608744c3c06302bf81d174')),
-        'amount': 35000000000000000000,
-        'balance': 2000,
-    },
+    'message': None
 }
 
 def main(accounts):
     from py_eth_sig_utils import signing
 
-    data = data2
+    data['message'] = {
+        'channel_adr': '0x254dffcd3277C0b1660F6d42EFbB754edaBAbC2B',
+        'channel_seq': 39,
+        'balance': 2700,
+    }
+    # signature: 0xe32976b152f5d3107a789bee8512741493c262984145415c1ffb3a42c1a80e7224dd52cc552bf86665dd185d9e04004eb8d783f624eeb6aab0011c21757e6bb21b
 
     # generate a new raw random private key
     if True:
+        # maker_key
+        # pkey_raw = a2b_hex('6370fd033278c143179d81c5526140625662b8daa446c22ee2d73db3707e620c')
+
+        # consumer_delegate_key
+        #pkey_raw = a2b_hex('e485d098507f54e7733a205420dfddbe58db035fa577fc294ebd14db90767a52')
         pkey_raw = a2b_hex('a4985a2ed93107886e9a1f12c7b8e2e351cc1d26c42f3aab7f220f3a7d08fda6')
     else:
         pkey_raw = os.urandom(32)
@@ -125,13 +111,21 @@ def main(accounts):
     caddr = web3.Web3.toChecksumAddress(addr)
     print('Account canonical address: {}'.format(caddr))
 
+    # step-wise computation of signature
+    msg_hash = signing.encode_typed_data(data)
+    print('Ok, MSG_HASH = 0x{}'.format(b2a_hex(msg_hash).decode()))
+    sig_vrs = utils.ecsign(msg_hash, pkey_raw)
+    sig = signing.v_r_s_to_signature(*sig_vrs)
+
     signature = signing.v_r_s_to_signature(*signing.sign_typed_data(data, pkey_raw))
     assert len(signature) == 32 + 32 + 1
-    print('Ok, signed typed data using {}:\nSIGNATURE = 0x{}'.format(caddr, b2a_hex(signature).decode()))
+    #assert signature == sig
+    print('Ok, signed typed data (using key {}):\nSIGNATURE = 0x{}'.format(caddr, b2a_hex(signature).decode()))
 
     signer_address = signing.recover_typed_data(data, *signing.signature_to_v_r_s(signature))
     assert signer_address == caddr
     print('Ok, verified signature was signed by {}'.format(signer_address))
+
 
 
 if __name__ == '__main__':
@@ -158,7 +152,7 @@ if __name__ == '__main__':
         print('could not connect to Web3/Ethereum at "{}"'.format(args.gateway or 'auto'))
         sys.exit(1)
     else:
-        print('connected to network {} at provider "{}"'.format(w3.version.network, args.gateway or 'auto'))
+        print('connected via provider "{}"'.format(args.gateway or 'auto'))
 
     # set new provider on XBR library
     xbr.setProvider(w3)